requirements.txt
# LinkedIn Automation - Requirements
# Install with: pip install -r requirements.txt

# Core
selenium>=4.10.0
webdriver-manager>=4.0.0
Pillow>=10.0.0

# Image handling
requests>=2.31.0

# Database (built-in sqlite3, no install needed)

# For GUI (tkinter is built-in with Python)